+++ docs/source/model_doc/transformers/whisper.mdx
@@ -121,4 +121,4 @@ pipeline("https://huggingface.co/datasets/Narsil/asr_dummy/resolve/main/mlk.flac
 
 ## NeuronWhisperForConditionalGeneration
 
-[[autodoc]] modeling_seq2seq.NeuronWhisperForConditionalGeneration
+[[autodoc]] models.whisper.NeuronWhisperForConditionalGeneration
