+++ optimum/neuron/models/whisper/__init__.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-# Copyright 2024 The HuggingFace Team. All rights reserved.
+# Copyright 2025 The HuggingFace Team. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,11 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from . import (
-    bert,
-    granite,
-    llama,
-    phi3,
-    qwen2,
-    whisper,
-)
+from .model import NeuronWhisperForConditionalGeneration
