+++ optimum/neuron/utils/__init__.py
@@ -125,6 +125,11 @@ _import_structure = {
         "NEURON_CAUSALLM_INPUTS_DOCSTRING",
         "NEURON_MODEL_START_DOCSTRING",
         "NEURON_CAUSALLM_MODEL_START_DOCSTRING",
+        "NEURON_TRANSLATION_EXAMPLE",
+        "NEURON_TRANSLATION_TP_EXAMPLE",
+        "NEURON_SEQ2SEQ_INPUTS_DOCSTRING",
+        "NEURON_AUDIO_SEQ2SEQ_INPUTS_DOCSTRING",
+        "NEURON_SEQ2SEQ_MODEL_START_DOCSTRING",
         "add_start_docstrings",
         "add_start_docstrings_to_model_forward",
     ],
@@ -161,6 +166,7 @@ if TYPE_CHECKING:
         NEURON_AUDIO_CLASSIFICATION_PIPELINE_EXAMPLE,
         NEURON_AUDIO_FRAME_CLASSIFICATION_EXAMPLE,
         NEURON_AUDIO_INPUTS_DOCSTRING,
+        NEURON_AUDIO_SEQ2SEQ_INPUTS_DOCSTRING,
         NEURON_AUDIO_XVECTOR_EXAMPLE,
         NEURON_CAUSALLM_INPUTS_DOCSTRING,
         NEURON_CAUSALLM_MODEL_START_DOCSTRING,
@@ -179,10 +185,14 @@ if TYPE_CHECKING:
         NEURON_SEMANTIC_SEGMENTATION_EXAMPLE,
         NEURON_SEMANTIC_SEGMENTATION_PIPELINE_EXAMPLE,
         NEURON_SENTENCE_TRANSFORMERS_EXAMPLE,
+        NEURON_SEQ2SEQ_INPUTS_DOCSTRING,
+        NEURON_SEQ2SEQ_MODEL_START_DOCSTRING,
         NEURON_SEQUENCE_CLASSIFICATION_EXAMPLE,
         NEURON_TEXT_GENERATION_EXAMPLE,
         NEURON_TEXT_INPUTS_DOCSTRING,
         NEURON_TOKEN_CLASSIFICATION_EXAMPLE,
+        NEURON_TRANSLATION_EXAMPLE,
+        NEURON_TRANSLATION_TP_EXAMPLE,
         add_start_docstrings,
         add_start_docstrings_to_model_forward,
     )
